Fix : 파일 분리
각각의 타입들을 Type.swift라는 파일에 따로 저장해서 main.swift에서는 코드 실행부만 존재하게 만들었습니다!
